pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agri-vlm-v1"
version = "0.1.0"
description = "Research codebase for agriculture-specialized vision-language model training and evaluation."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
  {name = "OpenAI Codex"}
]
dependencies = [
  "accelerate>=1.10.1,<2",
  "bitsandbytes>=0.48.1,<0.49",
  "datasets>=4.4.0,<5",
  "huggingface_hub>=0.35.0,<1",
  "numpy>=1.26,<3",
  "packaging>=24.2,<25",
  "pandas>=2.2,<3",
  "peft>=0.18.0,<0.19",
  "Pillow>=10.4,<12",
  "pydantic>=2.11.0,<3",
  "PyYAML>=6.0.2,<7",
  "scikit-learn>=1.6,<2",
  "torch>=2.6,<3",
  "transformers>=5.5.0,<5.6.0",
  "trl>=0.25.0,<0.26.0",
  "typing_extensions>=4.12,<5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-cov>=6.1,<7"
]
flash-attn = [
  "flash-attn>=2.8.0.post2"
]
deepspeed = [
  "deepspeed>=0.17.4,<0.18"
]
vllm = [
  "vllm>=0.11.0,<0.12"
]
qwen-utils = [
  "qwen-vl-utils>=0.0.14,<0.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
